+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - Use '\0' convention everywhere for empty zero terminated strings
+- fromStr(): Empty input string results empty host object
 
 ## [0.4.0]